refactor: decouple express app and index.ts
